Test coverage datastore 100%.
