+++ test/global/test_exceptions.py
