+++ test/global/test_logging.py
